third_party/blink/web_tests/editing/inserting/insert-newline-into-dynamically-created-textarea-expected.txt

PASS visibleNewline(textarea.value) is "hello//world"
PASS successfullyParsed is true

TEST COMPLETE
